requirements.txt
filelock==3.20.0
fsspec==2025.9.0
Jinja2==3.1.6
MarkupSafe==3.0.3
mpmath==1.3.0
networkx==3.5
sympy==1.14.0
torch==2.3.1+cpu
typing_extensions==4.15.0
fastapi==0.119.0
pillow==11.3.0
pydantic==2.12.2
pydantic_core==2.41.4
python-dotenv==1.1.1
python-multipart==0.0.20
rich==14.2.0
typer==0.19.2
uvicorn==0.37.0
watchfiles==1.1.1
numpy==2.3.3
